added Autocomplete and DropDownContent components also refactored the code
