+++ src/App.tsx
@@ -1,43 +1,12 @@
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import './App.scss';
-import debounce from 'lodash.debounce';
-import { peopleFromServer } from './data/people';
-import { DropDownMenu } from './components/DropDownMenu/DropDownMenu';
 import { Person } from './types/Person';
+import { Autocomplete } from './components/Autocomplete/Autocomplete';
 
 export const App: React.FC = () => {
-  const [query, setQuery] = useState('');
   const [selectedPerson, setSelectedPerson] = useState<Person | null>(null);
-  const [isInputFocused, setInputFocused] = useState(false);
-  const [appliedQuery, setAppliedQuery] = useState('');
 
-  const handleInputFocus = () => {
-    setInputFocused(true);
-  };
-
-  const applyQuery = useCallback(
-    debounce(setAppliedQuery, 1000),
-    [],
-  );
-
-  const handleQueryChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setQuery(event.target.value);
-    applyQuery(event.target.value);
-  };
-
-  const handlePersonSelect = useCallback((person: Person) => {
-    setSelectedPerson(person);
-    setQuery(person.name);
-    setAppliedQuery(person.name);
-    setInputFocused(false);
-  }, []);
-
-  const filteredPersons = useMemo(() => {
-    return peopleFromServer.filter(person => person.name.toLowerCase()
-      .includes(appliedQuery.toLowerCase()));
-  }, [appliedQuery]);
-
-  // const { name, born, died } = peopleFromServer[0];
+  const DEBOUNCE_TIME = 300;
 
   return (
     <main className="section">
@@ -47,25 +16,10 @@ export const App: React.FC = () => {
           : 'No selected person'}
       </h1>
 
-      <div className="dropdown is-active">
-        <div className="dropdown-trigger">
-          <input
-            type="text"
-            placeholder="Enter a part of the name"
-            className="input"
-            value={query}
-            onChange={handleQueryChange}
-            onFocus={handleInputFocus}
-          />
-        </div>
-
-        {isInputFocused && (
-          <DropDownMenu
-            people={filteredPersons}
-            onSelect={handlePersonSelect}
-          />
-        )}
-      </div>
+      <Autocomplete
+        setSelectedPerson={setSelectedPerson}
+        DEBOUNCE_TIME={DEBOUNCE_TIME}
+      />
     </main>
   );
 };
